Move previous iterator stuff into previous.rs
