+++ crates/fs/private/mod.rs
@@ -5,6 +5,7 @@ pub mod hamt;
 mod key;
 pub mod namefilter;
 mod node;
+mod previous;
 mod rng;
 
 pub use directory::*;
@@ -12,4 +13,5 @@ pub use file::*;
 pub use forest::*;
 pub use key::*;
 pub use node::*;
+pub use previous::*;
 pub use rng::*;
